--- src/NeoSharp.BinarySerialization/IFixedBufferConverter.cs
@@ -1,10 +1,0 @@
-﻿namespace NeoSharp.BinarySerialization
-{
-    public interface IFixedBufferConverter
-    {
-        /// <summary>
-        /// Fixed length of the buffer
-        /// </summary>
-        int FixedLength { get; }
-    }
-}
